backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-pptx==1.0.2
anthropic==0.40.0
pdfplumber==0.11.0
matplotlib==3.9.0
Pillow==10.4.0
pydantic==2.8.0
python-multipart==0.0.9
